feat: add show_user tool for terminal markdown rendering
Implement a new tool that renders markdown content in the terminal with proper formatting using termimad library. This includes:

- Adding termimad dependency to workspace and forge_app crate
- Creating new ToolRegistry struct to centralize tool registration
- Implementing show_user tool with full markdown rendering capabilities
- Updating tool_service to use the new registry
- Adding documentation in docs/show-user-tool.md
