+++ Cargo.toml
@@ -61,6 +61,7 @@ strum_macros = "0.27.1"
 syn = { version = "2.0.98", features = ["full"] }
 sysinfo = "0.33.1"
 tempfile = "3.10.1"
+termimad = "0.31.2"
 thiserror = "2.0.11"
 tokio = { version = "1.43.0", features = ["full", "test-util"] }
 tokio-stream = "0.1.17"
