+++ crates/forge_app/Cargo.toml
@@ -48,6 +48,7 @@ base64.workspace = true
 strum_macros.workspace = true
 strum.workspace = true
 bytes.workspace = true
+termimad.workspace = true
 
 [dev-dependencies]
 insta.workspace = true
